Found the real predict_loading_concentration
The real predict_loading_concentration function that was called in the two predict_loading_concentration.py files was located in __init__.py.
This real function was moved to diva_seq_opt/utility.
The duplicated predict_loading_concentration.py in diva_seq_opt was deleted, since the same actions are done by the parent directory's predict_loading_concentration.py.
The __init__.py was cleared of all text, so it serves only to make diva_seq_opt a module.
The main predict_loading_concentration.py was modified to make the call to predict_loading_concentration() in utility.py.